created HOC to fetch all staticContent, updated tests
